Update base.js
executing light commands in persistent mode seems to cause damage to some magic home devices. this caused my device to stop working after playing with the light intensity for a few minutes. by forcing flux_led to work in volatile mode, it started to work again. A little bit of googling yielded this:
https://pbxinaflash.com/community/threads/onair-light.19085/
apparently this can quickly damage the nvram of the device. fortunately, the device still works as long as i don't try to persist the updates. i'm proposing that flux_led always be called in volatile mode